Transfered update loop logic from view to viewmodel so that it can be interrupted by a stage closing handler in the injected scope.
